% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create dataset.}
\usage{
create_dataset(dataset, dataset_name, repo_id, alphacast_api_key)
}
\arguments{
\item{dataset}{Path of the CSV you want to upload to Alphacast.}

\item{dataset_name}{Name of the dataset.}

\item{repo_id}{Must be numeric. This is the id from the repo you want to upload the CSV.}

\item{alphacast_api_key}{API Key from Alphacast. You can get it from your Settings menu on the Alphacast Web.}
}
\value{
A dataset containing all the "lucky words" in the description.
}
\description{
Easily upload a CSV file to the Alphacast platform.
}
\examples{
alphacast_api_key = "ak_QDPD89fxwASlhWwdfeIO"
lucky_words <- c("mobility", "apple")
df <- get_lucky_dataset(lucky_words, alphacast_api_key, long = TRUE)
}
