% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repo.R
\name{create_repo}
\alias{create_repo}
\title{Create repos.}
\usage{
create_repo(repo_name, description, private = TRUE, slug, alphacast_api_key)
}
\arguments{
\item{repo_name}{Name of the repo you want to create.}

\item{description}{Short description of the new repo.}

\item{private}{Logical. If FALSE, the repo is public. If TRUE, the repo is private. Default: TRUE.}

\item{slug}{Short name of the repo. It can't contain spaces or mayus. It is going to be a part of the repo's URL. Example: my-private-repo.}

\item{alphacast_api_key}{API Key from Alphacast. You can get it from your Settings menu on the Alphacast Web.}
}
\value{
Creates a repo.
}
\description{
Create a repository in Alphacast environment with your API Key.
}
\examples{
alphacast_api_key = "ak_QDPD89fxwASlhWwdfeIO"

}
