% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets index}
\usage{
get_datasets(alphacast_api_key)
}
\arguments{
\item{alphacast_api_key}{API Key from Alphacast. You can get it from your Settings menu on the Alphacast Web.}
}
\value{
A index with all available datasets.
}
\description{
Get all available datasets from Alphacast with your level of access.
}
\examples{
alphacast_api_key = "ak_QDPD89fxwASlhWwdfeIO"
datasets_alphacast <- get_datasets(alphacast_api_key)
}
