% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lucky_dataset.R
\name{get_lucky_dataset}
\alias{get_lucky_dataset}
\title{Get Lucky Dataset}
\usage{
get_lucky_dataset(lucky_words, alphacast_api_key, long = FALSE)
}
\arguments{
\item{lucky_words}{One or more words in a vector. It can be a country, a variable or whatever!}

\item{alphacast_api_key}{API Key from Alphacast. You can get it from your Settings menu on the Alphacast Web.}

\item{long}{Logical. Choose between data format.}
}
\value{
A dataset containing all the "lucky words" in the description.
}
\description{
¿Feelin' lucky? Get a random dataset containing one of two words you can choose to filter.
}
\examples{
alphacast_api_key = "ak_QDPD89fxwASlhWwdfeIO"
lucky_words <- c("mobility", "apple")
df <- get_lucky_dataset(lucky_words, alphacast_api_key, long = TRUE)
}
