% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repos.R
\name{get_repos}
\alias{get_repos}
\title{Get repositories}
\usage{
get_repos(alphacast_api_key)
}
\arguments{
\item{alphacast_api_key}{API Key from Alphacast. You can get it from your Settings menu on the Alphacast Web.}
}
\value{
A index with all available repos.
}
\description{
Get all available repositories from Alphacast with your level of access.
}
\examples{
alphacast_api_key = "ak_QDPD89fxwASlhWwdfeIO"
repos_alphacast <- get_repos(api_key)
}
